requirement.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
ethiopian-date-converter==0.1.5
importlib-metadata==6.0.0
Markdown==3.4.1
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
zipp==3.12.1
